requirements.txt
httpx==0.27.0
uvicorn==0.30.5
oracledb~=3.4.0
PyJWT~=2.10.1
fastapi==0.112.0
python-dotenv==1.0.1
slowapi==0.1.9
passlib==1.7.4
jwcrypto~=1.5.6
cryptography==3.4.8